requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.58.2
django-phonenumber-field==7.2.0
dnspython==2.4.2
idna==3.4
mongoengine==0.27.0
oauthlib==3.2.2
phonenumbers==8.13.27
phonenumberslite==8.13.27
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.5.0
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.1.0
